+++ packages/authentication/modules/access/useLogin/__tests__/useLogin.test.tsx
@@ -6,6 +6,8 @@ import {
 } from '@baseapp-frontend/test'
 import { axios } from '@baseapp-frontend/utils'
 
+import { z } from 'zod'
+
 import useLogin from '../index'
 
 // @ts-ignore TODO: (BA-1081) investigate AxiosRequestHeaders error
@@ -45,5 +47,45 @@ describe('useLogin', () => {
 
     expect(hasOnSuccessRan).toBe(true)
     expect(cookiesMock.set).toBeCalledTimes(2)
+  })
+
+  test('should allow custom defaultValues and validationSchema', async () => {
+    axiosMock.onPost('/auth/login').reply(200, {})
+
+    const customDefaultValues = {
+      email: 'test@tsl.io',
+      phoneNumber: '12345',
+      password: 'fW7q0jwv',
+    }
+    const customValidationSchema = z.object({
+      password: z.string().nonempty(),
+      phoneNumber: z
+        .string()
+        .nonempty()
+        .regex(/^\d{5}$/),
+      email: z.string().nonempty().email(),
+    })
+
+    let hasOnSuccessRan = false
+
+    const { result } = renderHook(
+      () =>
+        useLogin({
+          defaultValues: customDefaultValues,
+          validationSchema: customValidationSchema,
+          loginOptions: {
+            onSuccess: () => {
+              hasOnSuccessRan = true
+            },
+          },
+        }),
+      {
+        wrapper: ComponentWithProviders,
+      },
+    )
+
+    await result.current.form.handleSubmit()
+
+    expect(hasOnSuccessRan).toBe(true)
   })
 })