+++ packages/authentication/modules/mfa/useMfaActivateConfirm/types.ts
@@ -1,4 +1,5 @@
 import { UseMutationOptions } from '@tanstack/react-query'
+import { z } from 'zod'
 
 import MfaApi from '../../../services/mfa'
 import { IMfaConfirmationResponse, IMfaRequest, MfaMethod } from '../../../types/mfa'
@@ -6,8 +7,7 @@ import { IMfaConfirmationResponse, IMfaRequest, MfaMethod } from '../../../types
 type ApiClass = Pick<typeof MfaApi, 'confirmActivation'>
 
 export interface IUseMfaActivateConfirm {
-  // TODO: refactor types
-  validationSchema?: any
+  validationSchema?: z.ZodObject<z.ZodRawShape>
   defaultValues?: Partial<IMfaRequest>
   options?: UseMutationOptions<IMfaConfirmationResponse, unknown, IMfaRequest, any>
   ApiClass?: ApiClass