+++ packages/utils/constants/yup.ts
@@ -1,1 +1,0 @@
-export const YUP_REQUIRED_FIELD = 'This field is required.'