+++ packages/utils/constants/zod.ts
@@ -1,0 +1,7 @@
+export const ZOD_MESSAGE = {
+  required: 'This field is required.',
+  email: 'Please provide a properly formatted email address.',
+  passwordDoNotMatch: 'Passwords do not match.',
+  phoneNumber: 'Please provide a properly formatted phone number.',
+  mustAcceptTerms: 'Must Accept Terms and Conditions.',
+}